fix(volo-cli): only print err msg once
In previous codes, the `error` macro will print error messages, and
the returned value of `main` function will print the same content
again.

To avoid the problem, just remove the returned value for `main` and
exit the process by itself.

In the previous code, the `error` macro will print error message, and
the return value of the `main` function will print the same content
again.

To avoid this problem, just remove the return value of `main` and exit.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>