feat(volo-http): refactor `HttpContext` with `RpcCx`
This commit deletes the original `HttpContext` and creates a new type
`ServerContext` using `RpcCx`. The current `HttpContext` is just an
alias of `ServerContext`.

Because `ServerContext`, `RpcCx`, `ServerCxInner` all implement
`Deref` and `DerefMut`, no breaking changes are expected.

Please note that `HttpContext` may be removed in the future, so it is
recommended to use `ServerContext`.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>