chore(volo-http): remove state support
`State` is very intrusive and makes the implementation of `Route` and
`Handler` complicated, and the feature `State` can be replaced by
`lazy_static` or `Extension`.

Due to the above reasons, we think `State` is not necessary, so this
commit removes it in internal implementation of `volo-http`.

Because the `with_state` was previously disabled, and we did not
remove `state` in `FromContext` and `FromRequest`, no breaking changes
were introduced.

The `state` in `FromContext` and `FromRequest` will be completely
removed in the future.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>