chore(volo): use unix `SocketAddr` for unix socket addr (#353)
* chore(volo): use unix `SocketAddr` for unix socket addr

The previous `Address` used a path to a unix socket address, but it
did not support unnamed socket addresses, such as the incoming
address.

To support incoming unix socket addresses, we replace the path in
`Address::Unix` with `std::os::unix::net::SocketAddr`.

We also noticed that `tokio` uses a self-written `SocketAddr` because
it cannot create a `SocketAddr` from `std`. As a unix socket address,
their `struct` is the same, so we can use `std::mem::transmute` to
convert them to one another. Because `std::mem::transmute` guarantees
that both types have the same size, we can convert it directly and
safely.

* chore(ci): remove beta toolchains

---------

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>