bugfix(volo-http): add unit tests for client and fix some bugs
This commit adds some unit tests for common client functions, and
fixes some bugs during testing them.

The `hickory_resolver::Resolver` is a sync resolver, and
`Resolver::resolve` is a sync function, but it uses
`Runtime::block_on` of `tokio::runtime::Runtime`, which will **panic**
when it is in the scope of a runtime.

This commit removed dependence for sync resolver, which means that we
do not resolve host name when building client, instead of resolving it
when sending request, during which async resolver is available and
will not panic on `Runtime`.

And the config `disable_tls` should prevent sending HTTPS request, but
in the previous implement the client will send it as an HTTP request,
this commit also fixed it by returning an error `BadScheme`.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>