chore(volo-http): do not extract request in `from_fn`
Since #434 updated `from_fn` and it can modify types of request body
and error, but extractor of request will refuse any body without
`http_body::Body`.

Considering that extractor of request is not common when using
`from_fn`, we remove it in this commit, and add support for changing
type of response.

With this commit, users can build more complex service, and example
is also updated.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>