fix(volo-http): use `Option<CallOpt>` rather than `CallOpt`
In the previous code, if there was a default `Target` and the request
only had a `CallOpt` set, the new `CallOpt` would have no effect
because the request had no `Target`.

This commit fixes this by replacing `CallOpt` with `Option<CallOpt>`.
If we use the default target and there is only one `CallOpt`, we will
use it directly, otherwise the target `CallOpt` will be used. If the
request target is used, the default `CallOpt` will have no effect.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>