chore(volo-http): refactor `RequestPartsExt`
The previous implementation of `RequestPartsExt` was inelegant and
somewhat inaccurate, so we remove most functions in `RequestPartsExt`,
implement it for both `http::request::Parts` and
`http::request::Request`, move it from `volo_http::context::server` to
`volo_http::request`.

In addition, `RequestPartsExt::full_uri` has been changed to extractor
`FullUri` because it depends status of tls.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>