fix(volo-http): prefer ipv6 addr in dns resolver when dns addr is ipv6
We use the `hickory_resolver` crate to resolve domain names, but we
found that it always prefers IPv4 addresses, which doesn't work if the
client is running in an IPv6 only environment.

This commit fixes this by checking the first name server, if the
address is an IPv4 address we keep preferring IPv4 addresses, if it is
an IPv6 address we set the resolver to prefer IPv6 addresses.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>