fix(volo-http): fix `clippy::result_large_err`
The type `Uri` takes up 88 bytes, which causes the entire `ClientError`
to take up 144 bytes and clippy will throw `clippy::result_large_err`.

Considering that the field will not be used in the hot path, in order
to avoid the problem of `ClientError` being too large, we added `Box`
to `Uri`.

With this change, the size of `ClientError` is reduced from 144 to 64.

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>