feat(volo-http): support http2 and connection pool for client (#574)
This commit adds HTTP/2 support for the HTTP client and connection
pool support for both HTTP/1 and HTTP/2.

As the client interface is relatively simple, this commit introduces a
small number of breaking changes:

Since this PR supports HTTP/2, some configurations that only apply to
HTTP/1 are marked as `deprecated` or removed. Users should configure
via `builder.http1_config()`.

TODO:

1. `rustls` supports setting ALPN for each TLS connection, but
   `tokio-rustls` does not expose this feature, it can only set ALPN
   for the entire configuration (in other words,
   `tokio_rustls::TlsConnector`). We can use `rustls` directly or
   submit an issue to `tokio-rustls` for discussion to fix it.
2. `httpbin` supports HTTP/1 based on pure TCP or TLS, and HTTP/2
   based on TLS, but does not support HTTP/2 based on Cleartext (h2c),
   so there is currently no test case that only supports HTTP/2
   without TLS (although it is a unnecessary feature). We should find
   a site that supports h2c for subsequent testing.

---------

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>