fix(volo-http): put stripping IPv6 brackets in front.
In a previous commit, we fixed an issue where the DNS resolver didn't
strip brackets from IPv6 addresses
(it accepted `::1` instead of `[::1]`), but we only handled this
during DNS resolution and ignored the logic for SNI delivery, while
IPv6 addresses using TLS suffered from the same issue during the TLS
handshake.

This commit move the stripping logics from DNS resolving to service
name setting so that it works on everywhere.

Fixes: #597

Signed-off-by: Yu Li <liyu.yukiteru@bytedance.com>